Project creation dialog is now a screen
